# contracts/openapi/api.yaml
openapi: 3.1.0
info:
  title: Dorahub Internal API
  version: 0.18.0
  description: |
    JSON uses lowerCamelCase property names and RFC 3339 UTC timestamps.
    Collection endpoints use zero-based pages and the shared page, size, sort and filter parameters.
servers:
  - url: /
paths:
...
  /api/v1/tables:
    get:
      operationId: listMyTables
      summary: Tables the authenticated player sits at, most recent first
      security:
        - sessionCookie: []
      parameters:
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 50, default: 20 }
        - $ref: "#/components/parameters/RequestCorrelationId"
        - $ref: "#/components/parameters/ClientSource"
      responses:
        "200":
          description: Tables of the player
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TableSummaryView"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createTable
      summary: Create a table in lobby state with an immutable ruleset snapshot
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/RequestCorrelationId"
        - $ref: "#/components/parameters/ClientSource"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTableRequest"
      responses:
        "201":
          description: Created table
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TableView"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /api/v1/tables/stats:
    get:
      operationId: getTableStats
      summary: Pilot metrics over tables; moderator only, no personal data
      parameters:
        - $ref: "#/components/parameters/RequestCorrelationId"
        - $ref: "#/components/parameters/ClientSource"
      responses:
        "200":
          description: Counts and durations over all tables
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TableStatsView"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/tables/hands/mine:
    get:
      operationId: getMyHands
      summary: Own won hands, newest first; reverted ones are left out
      parameters:
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
        - $ref: "#/components/parameters/RequestCorrelationId"
        - $ref: "#/components/parameters/ClientSource"
      responses:
        "200":
          description: Won hands of the signed-in player
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WonHandView"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/tables/{tableId}:
    get:
      operationId: getTable
      summary: Return the current table state and aggregate version
      parameters:
        - $ref: "#/components/parameters/TableId"
        - $ref: "#/components/parameters/RequestCorrelationId"
        - $ref: "#/components/parameters/ClientSource"
      responses:
        "200":
          description: Table state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TableView"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/tables/{tableId}/events:
    get:
      operationId: getTableEvents
      summary: Poll the append-only table journal after a known sequence
      parameters:
        - $ref: "#/components/parameters/TableId"
        - $ref: "#/components/parameters/EventsSince"
        - $ref: "#/components/parameters/RequestCorrelationId"
        - $ref: "#/components/parameters/ClientSource"
      responses:
        "200":
          description: Events ordered by sequence
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TableEventView"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/tables/{tableId}/players:
    post:
      operationId: joinTable
      summary: Join the lobby as the authenticated account
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/TableId"
        - $ref: "#/components/parameters/RequestCorrelationId"
        - $ref: "#/components/parameters/ClientSource"
      responses:
        "200":
          description: Table state after joining
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TableView"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/tables/{tableId}/players/me:
    delete:
      operationId: leaveLobby
      summary: Leave a lobby before the game starts
      parameters:
        - $ref: "#/components/parameters/TableId"
        - $ref: "#/components/parameters/RequestCorrelationId"
        - $ref: "#/components/parameters/ClientSource"
      responses:
        "200":
          description: Table state without the player
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TableView"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/tables/{tableId}/start:
    post:
      operationId: startTable
      summary: Seat the four participants and start the game
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/TableId"
        - $ref: "#/components/parameters/RequestCorrelationId"
        - $ref: "#/components/parameters/ClientSource"
      responses:
        "200":
          description: Active table
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TableView"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/tables/{tableId}/riichi:
    post:
      operationId: declareRiichi
      summary: Move a 1000 point stick from a seat onto the table
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/TableId"
        - $ref: "#/components/parameters/RequestCorrelationId"
        - $ref: "#/components/parameters/ClientSource"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RiichiRequest"
      responses:
        "200":
          description: Table state after the declaration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TableView"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/tables/{tableId}/hands/preview:
    post:
      operationId: previewHand
      summary: Score a hand without changing the table
      parameters:
        - $ref: "#/components/parameters/TableId"
        - $ref: "#/components/parameters/RequestCorrelationId"
        - $ref: "#/components/parameters/ClientSource"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HandRequest"
      responses:
        "200":
          description: Score breakdown and payments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HandPaymentView"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /api/v1/tables/{tableId}/hands:
    post:
      operationId: confirmHand
      summary: Apply a won hand; requires the aggregate version the preview was built on
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/TableId"
        - $ref: "#/components/parameters/RequestCorrelationId"
        - $ref: "#/components/parameters/ClientSource"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HandRequest"
      responses:
        "200":
          description: Applied payments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HandPaymentView"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /api/v1/tables/{tableId}/draws:
    post:
      operationId: recordDraw
      summary: Record an exhaustive or abortive draw
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/TableId"
        - $ref: "#/components/parameters/RequestCorrelationId"
        - $ref: "#/components/parameters/ClientSource"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DrawRequest"
      responses:
        "200":
          description: Table state after the draw
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TableView"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/tables/{tableId}/seats/{seat}/player:
    post:
      operationId: takeSeat
      summary: Take a seat left free mid-game; the game stops counting towards rating
      parameters:
        - $ref: "#/components/parameters/TableId"
        - $ref: "#/components/parameters/Seat"
        - $ref: "#/components/parameters/RequestCorrelationId"
        - $ref: "#/components/parameters/ClientSource"
      responses:
        "200":
          description: Table state with the seat taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TableView"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
    delete:
      operationId: leaveSeat
      summary: Leave the table mid-game; scores stay and the seat waits for a substitute
      parameters:
        - $ref: "#/components/parameters/TableId"
        - $ref: "#/components/parameters/Seat"
        - $ref: "#/components/parameters/RequestCorrelationId"
        - $ref: "#/components/parameters/ClientSource"
      responses:
        "200":
          description: Table state with the seat free
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TableView"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/tables/{tableId}/revert:
    post:
      operationId: revertTable
      summary: Return an active table to a version it previously had (moderator only)
      parameters:
        - $ref: "#/components/parameters/TableId"
        - $ref: "#/components/parameters/RequestCorrelationId"
        - $ref: "#/components/parameters/ClientSource"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RevertRequest"
      responses:
        "200":
          description: Table state after the revert
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TableView"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /api/v1/tables/{tableId}/finish:
    post:
      operationId: finishTable
      summary: Finish the table early and publish final standings
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/TableId"
        - $ref: "#/components/parameters/RequestCorrelationId"
        - $ref: "#/components/parameters/ClientSource"
      responses:
        "200":
          description: Finished table with standings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TableView"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: DORAHUB_SESSION
...
  parameters:
...
    RequestCorrelationId:
      name: X-Correlation-Id
      in: header
      description: Optional caller-generated request correlation identifier.
      schema:
        type: string
        pattern: "^[A-Za-z0-9._-]{1,64}$"
    ClientSource:
      name: X-Client-Source
      in: header
      description: Low-cardinality client channel, without device or user identifiers.
      schema:
        type: string
        pattern: "^[A-Za-z0-9._-]{1,32}$"
...
  responses:
    BadRequest:
      $ref: "#/components/responses/ProblemResponse"
      description: Malformed request (`request.invalid`).
    Unauthorized:
      $ref: "#/components/responses/ProblemResponse"
      description: Authentication is required (`authentication.required`).
...
    UnprocessableEntity:
      $ref: "#/components/responses/ProblemResponse"
      description: Request is syntactically valid but cannot be processed (`request.unprocessable`).
...
  schemas:
    CreateTableRequest:
      type: object
      required: [rulesetKey, format]
      properties:
        rulesetKey: { type: string, minLength: 1 }
        format:
          type: string
          enum: [HANCHAN, TONPUUSEN]
...
    TableSummaryView:
      type: object
      required: [id, state, format, handsPlayed, updatedAt]
      properties:
        id: { type: string, format: uuid }
        state:
          type: string
          enum: [lobby, active, finished]
        format:
          type: string
          enum: [hanchan, tonpuusen]
        handsPlayed: { type: integer, minimum: 0 }
        updatedAt: { type: string, format: date-time }
...
    TableView:
      type: object
      required:
        [id, state, format, rulesetId, participants, seats, scores, dealerSeat, roundWind,
         handNumber, honba, riichiSticks, handsPlayed, vacantSeats, version, nicknames,
         updatedAt]
      properties:
        id: { type: string, format: uuid }
        state:
          type: string
          enum: [lobby, active, finished]
        format:
          type: string
          enum: [hanchan, tonpuusen]
        rulesetId:
          type: string
          description: Ruleset key and version, for example "rrc-ru@1.0".
        participants:
          type: array
          items: { type: string, format: uuid }
        seats:
          type: array
          description: Players by seat once the game started; empty in lobby.
          items: { type: string, format: uuid }
        scores:
          type: array
          items: { type: integer }
        dealerSeat: { type: integer, minimum: 0, maximum: 3 }
        roundWind:
          type: string
          enum: [east, south, west, north]
        handNumber: { type: integer, minimum: 1, maximum: 4 }
        honba: { type: integer, minimum: 0 }
        riichiSticks: { type: integer, minimum: 0 }
        handsPlayed: { type: integer, minimum: 0 }
        vacantSeats:
          type: array
          description: >-
            Seats whose player left mid-game. Hands cannot be recorded until every seat is
            taken again.
          items: { type: integer, minimum: 0, maximum: 3 }
        version:
          type: integer
          format: int64
          description: Aggregate version; also the sequence of the last journal event.
        nicknames:
          type: object
          additionalProperties: { type: string }
          description: >-
            Nicknames of the participants by account id. A seat wind alone says nothing —
            players need to see who sits where.
        updatedAt:
          type: string
          format: date-time
          description: >-
            Last change of the table. A game with no new hands for a long time is shown as
            stale by the client; the server never closes a game on its own, only empty lobbies.
        finishedReason:
          type: string
          enum: [completed, early, abandoned_lobby]
          description: >-
            Why the table is finished; absent while the game is on. Only "completed" counts
            towards rating.
        standings:
          type: array
          description: Final places with uma; present only for a finished game.
          items:
            $ref: "#/components/schemas/PlacementView"
...
    TableStatsView:
      type: object
      required:
        [lobbies, active, stale, completed, abandonedEarly, abandonedLobby, completionRate,
         handsPerCompletedGame, medianMinutes, p90Minutes]
      properties:
        lobbies: { type: integer, minimum: 0 }
        active: { type: integer, minimum: 0 }
        stale:
          type: integer
          minimum: 0
          description: Active games with no new hand for over 12 hours.
        completed: { type: integer, minimum: 0 }
        abandonedEarly: { type: integer, minimum: 0 }
        abandonedLobby: { type: integer, minimum: 0 }
        completionRate:
          type: number
          description: Percentage of finished tables that were played to the end of the format.
        handsPerCompletedGame: { type: number }
        medianMinutes: { type: integer, format: int64 }
        p90Minutes: { type: integer, format: int64 }
...